Added the Virtual Console
The virtual console is an 8-bit console style project. It's still a work in progress, but it now has its own vaguely assembly like language to be programmed in. This also brings a bunch of bug fixes I can't be bothered to count, but the virtual console was the main addition.